+++ gulpfile.js
@@ -189,7 +189,7 @@ gulp.task('sync', (done) => {
   //Initialize Browsersync with a PHP server.
   connect.server({}, function () {
     browserSync.init(files, {
-      proxy: 'http://tylerpaulsonpictures.test/'
+      proxy: 'https://virtuoso.test/'
     });
 
   });
@@ -213,10 +213,10 @@ gulp.task('clean', (done) => {
 
 gulp.task('watch:code', () => {
   gulp.watch('./assets/sass/**/*.scss', gulp.series('styles'));
-  gulp.watch('./assets/js/src/*.js', gulp.series('js', reload) );
-  gulp.watch('./**/*.php').on('change', function() {
-    reload();
-  });
+  gulp.watch('./assets/js/src/*.js', gulp.series('js') );
+  // gulp.watch('./**/*.php').on('change', function() {
+  //   reload();
+  // });
 
 });
 
@@ -226,5 +226,5 @@ function reload() {
 
 gulp.task('build', gulp.series('clean', 'styles', 'js' ) );
 
-gulp.task('watch', gulp.series('build', 'sync', 'watch:code') );
+gulp.task('watch', gulp.series('build', 'watch:code') );